standardize task instructions for all ages
