+++ task-launcher/src/tasks/shared/helpers/index.ts
@@ -49,3 +49,4 @@ export { shouldTerminateCat } from './shouldTerminateCat';
 export { selectNextSequentialTrial } from './selectNextSequentialTrial';
 export { checkFallbackCriteria } from './checkFallbackCriteria';
 export { addKeyHelpers } from './addKeyHelpers';
+export { pulseOkButton } from './pulseOkButton';
